# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wealth Master API",
    "version": "0.8.0",
    "description": "Session authentication, mandatory production MFA, accounts, calculated cash balances, manual income/expenses/refunds, and same-currency paired transfers. Current net worth by native currency, including archived account cash balances. User-owned two-level income/spending categories and optional ledger assignment. Dated normal monthly/yearly limits, specific-period exceptions, native currency actuals and explicit exception copying. Spending distribution reports show expenses before refunds and exact net spending independently of limits."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/budgets/{id}/activity": {
      "get": {
        "summary": "Supporting expenses/refunds in ledger activity order",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              }
            }
          },
          "400": {
            "description": "Invalid decimal, currency, period or request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or MFA restrictions"
          },
          "404": {
            "description": "Missing, unowned or deleted budget/category"
          },
          "409": {
            "description": "Duplicate live limit or unavailable category branch"
          },
          "412": {
            "description": "Stale source/budget version; no writes committed"
          },
          "428": {
            "description": "Quoted If-Match required"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          }
        ],
        "description": "50 items ordered by transaction date descending, creation timestamp descending, ID; includes archived accounts/categories. Read-only repeatable snapshot."
      }
    },
    "/spending": {
      "get": {
        "summary": "Read spending distribution and selected-period comparisons",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpendingReport"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              }
            }
          },
          "400": {
            "description": "Invalid decimal, currency, period or request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or MFA restrictions"
          }
        },
        "parameters": [
          {
            "name": "periodType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "MONTH",
                "YEAR"
              ]
            },
            "description": "Defaults to MONTH"
          },
          {
            "name": "periodStart",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Canonical first day; defaults to current business period"
          }
        ],
        "description": "One read-only repeatable-read snapshot, owner-scoped and no-store. Includes every spending category and archived activity; excludes deleted operations, income, transfers and opening balances. Gross expenses and refunds remain separate; net spending drives limits. Direct amounts belong to the assigned category; inclusive amounts add immediate children. Pie uses root inclusive expenses plus uncategorized, counting each expense once. No budgets are required. Effective comparisons include dated defaults, period exceptions and accrued monthly totals; usage dates and business date are explicit. Activity is capped at the business date."
      }
    },
    "/spending/activity": {
      "get": {
        "summary": "Read spending activity with or without a limit",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityPage"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              }
            }
          },
          "400": {
            "description": "Invalid decimal, currency, period or request"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "CSRF or MFA restrictions"
          },
          "404": {
            "description": "Missing or unowned spending category"
          }
        },
        "parameters": [
          {
            "name": "periodType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "MONTH",
                "YEAR"
              ]
            },
            "description": "Defaults to MONTH"
          },
          {
            "name": "periodStart",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Canonical first day; defaults to current business period"
          },
          {
            "name": "currency",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Z]{3}$"
            },
            "description": "ISO 4217 currency"
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Owned spending category, omitted for uncategorized"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 1000000,
              "default": 0
            }
          }
        ],
        "description": "Owner-scoped repeatable-read snapshot; 50 operations per page in ledger order (transaction date descending, created timestamp descending, ID). Category includes direct assignments and immediate children. Omit categoryId for uncategorized. Archived activity is included. No limit is required."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      }
    },
    "schemas": {
...
      "LedgerOperation": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "INCOME",
              "EXPENSE",
              "REFUND",
              "TRANSFER"
            ]
          },
          "amount": {
            "type": "string",
            "pattern": "[0-9]{1,20}(\\.[0-9]{1,8})?",
            "description": "Strictly positive decimal; no rounding."
          },
          "transactionDate": {
            "type": "string",
            "format": "date"
          },
          "valueDate": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "payee": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "destinationAccountId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "currency": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "categoryId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryAssignment"
          }
        },
        "required": [
          "categoryId",
          "category"
        ]
      },
      "ActivityPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerOperation"
            }
          },
          "page": {
            "type": "integer"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "parentId",
          "active",
          "createdAt",
          "version",
          "hasActivity",
          "hasChildren",
          "available"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Trimmed; unique case-insensitively within owner/type/parent, including archived categories."
          },
          "type": {
            "type": "string",
            "enum": [
              "INCOME",
              "SPENDING"
            ]
          },
          "parentId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "hasActivity": {
            "type": "boolean",
            "description": "Permanent ledger or budget references, including deleted budgets/activity and previously selected parents."
          },
          "hasChildren": {
            "type": "boolean"
          },
          "available": {
            "type": "boolean",
            "description": "Category and parent are active."
          }
        }
      },
...
      "CategoryAssignment": {
        "oneOf": [
          {
            "type": "null"
          },
          {
            "type": "object",
            "required": [
              "id",
              "name",
              "parentName",
              "available"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "parentName": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "available": {
                "type": "boolean"
              }
            }
          }
        ]
      },
...
      "SpendingAmounts": {
        "type": "object",
        "required": [
          "expenses",
          "refunds",
          "netSpending"
        ],
        "properties": {
          "expenses": {
            "type": "string",
            "description": "Exact decimal amount; no floating point rounding"
          },
          "refunds": {
            "type": "string",
            "description": "Exact decimal amount; no floating point rounding"
          },
          "netSpending": {
            "type": "string",
            "description": "Exact decimal amount; no floating point rounding"
          }
        }
      },
      "SpendingGroup": {
        "type": "object",
        "required": [
          "categoryId",
          "currency",
          "direct",
          "inclusive"
        ],
        "properties": {
          "categoryId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Null denotes uncategorized"
          },
          "currency": {
            "type": "string"
          },
          "direct": {
            "$ref": "#/components/schemas/SpendingAmounts"
          },
          "inclusive": {
            "$ref": "#/components/schemas/SpendingAmounts"
          }
        }
      },
      "SpendingCurrency": {
        "type": "object",
        "required": [
          "currency",
          "expenses",
          "refunds",
          "netSpending",
          "uncategorized",
          "unbudgeted"
        ],
        "properties": {
          "currency": {
            "type": "string"
          },
          "expenses": {
            "type": "string"
          },
          "refunds": {
            "type": "string"
          },
          "netSpending": {
            "type": "string"
          },
          "uncategorized": {
            "type": "string"
          },
          "unbudgeted": {
            "type": "string"
          }
        }
      },
      "SpendingReport": {
        "type": "object",
        "required": [
          "periodType",
          "periodStart",
          "categories",
          "groups",
          "currencies",
          "limits",
          "businessDate",
          "settings"
        ],
        "properties": {
          "periodType": {
            "type": "string",
            "enum": [
              "MONTH",
              "YEAR"
            ]
          },
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpendingGroup"
            }
          },
          "currencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpendingCurrency"
            }
          },
          "limits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetComparison"
            }
          },
          "businessDate": {
            "type": "string",
            "format": "date"
          },
          "settings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BudgetSetting"
            }
          }
        }
      },
      "BudgetReference": {
        "type": "object",
        "required": [
          "id",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BudgetSetting": {
        "type": "object",
        "required": [
          "id",
          "categoryId",
          "currency",
          "mode",
          "limit",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "currency": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "NONE",
              "MONTH",
              "YEAR"
            ]
          },
          "limit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?[0-9]+(\\.[0-9]+)?$"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "BudgetComparison": {
        "type": "object",
        "required": [
          "categoryId",
          "currency",
          "periodType",
          "periodStart",
          "source",
          "limit",
          "actual",
          "remaining",
          "percentage",
          "overBudget",
          "usageStart",
          "usageEnd",
          "currentPeriod",
          "available",
          "settingReference",
          "overrideReference",
          "accruedMonths",
          "coveredMonths"
        ],
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "currency": {
            "type": "string"
          },
          "periodType": {
            "type": "string",
            "enum": [
              "MONTH",
              "YEAR"
            ]
          },
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "source": {
            "type": "string",
            "enum": [
              "NORMAL",
              "EXCEPTION",
              "MONTHLY_ROLLUP"
            ]
          },
          "limit": {
            "type": "string",
            "pattern": "^-?[0-9]+(\\.[0-9]+)?$"
          },
          "actual": {
            "type": "string",
            "pattern": "^-?[0-9]+(\\.[0-9]+)?$"
          },
          "remaining": {
            "type": "string",
            "pattern": "^-?[0-9]+(\\.[0-9]+)?$"
          },
          "percentage": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?[0-9]+(\\.[0-9]+)?$"
              },
              {
                "type": "null"
              }
            ]
          },
          "overBudget": {
            "type": "boolean"
          },
          "usageStart": {
            "type": "string",
            "format": "date"
          },
          "usageEnd": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ]
          },
          "currentPeriod": {
            "type": "boolean"
          },
          "available": {
            "type": "boolean"
          },
          "settingReference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "overrideReference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetReference"
              },
              {
                "type": "null"
              }
            ]
          },
          "accruedMonths": {
            "type": "integer"
          },
          "coveredMonths": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}